[WIP] [Feature] CRUD Order

+++ domain/order.go
@@ -4,6 +4,6 @@ package domain
 type Order struct {
 	Model
 	TableID      UUID          `json:"table_id"`
-	TotalPayment float64       `json:"total_payment"`
+	TotalPayment int           `json:"total_payment"`
 	OrderDetails []OrderDetail `json:"order_details"`
 }
